First pass at diff method
Method is based off one CondensedForulmae object and diffs to another. Diff is pretty crude at the moment and is
crying out for the finding of diffs to be de-coupled from the reporting of diffs, which is a future exercise.

This diff will group blocks of formulae together though so this is the first point at which this code is better than
na-ka-na's (in any aspect!).

Testing is pretty sparse so far so may need some work to iron out bugs.

Have not got close to implementing any searches for transformations, yet.