Add distance measure to Formula
This is a necessary step towards determining which rows and columns have
been moved around when trying to detect transalations.

I have noticed that when rows / columns are inserted / deleted then the
underlying formula structure will remain the same and all that's moved
will be the cell references. Have put a number measure on how much they've
moved as the preference will be to not move rows very much so need to
quantify how much movement has happened.