Create get row & column subsetting methods
These two new methods will take a CondensedFormulae object and return a new
CondensedFormulae object subsetted down to a single row or column. This is
necessary prep work for the row and column translation detection that is
coming.

In order to do this have had to go back into Formula and CellRef creating
new methods to construct the Formula that would apply to the first cell in
the subsetted block. This is because only one Formula is stored for each
unique formula on a sheet and that formula is specific to a certain cell
location. If the cell location isn't in the subsetted target then the
Formula is no longer appropriate for the subset created so we need to move
it. This is equivalent to copying for source cell to 1st subset cell