First pass at mapping rows and columns
This doesn't work yet but is substantively there and now needs debugging.
At the moment we're aiming at creating a map of which rows map to which
rows in a pair of CondensedFormulae objects. The exact same process is
sued for columns too. The idea is that with the map the various insert,
delete, move translations and subsequently be inferred