Only need to map in one direction
Having gone to all that trouble to be able to map in either direction I
don't think we need to map in both directions. This is because when you
loop through all the FROM rows and map them, either to no row at all on
the TO side or a row on the TO side then by definition you've searched
every row on the TO side as well. Either you've found a To match, and it
will be mapped to its FROM counterpart or there's no match to find, so it
won't be mapped.

For now have just commented out as am not 100% sure but will edit out
subsequently as I become more convinced in further testing