Test file with row and col insert
Now runs a map test against a file with a row and col insert. In this most
trivial of file comparisons it seems to work perfectly. Need to test on
more complex files but this is a major milestone

The same test the other way (i.e. with a row and col delete) as the delete
in both dimensions means the row / column in the from file can never be
fully found in the to file. Clearly this is to be fixed