Improve boundary safety
Was getting instances where the index went out of bounds (the positive
index started negative).

It looks like this was caused by the offset variable in mapAToB which is
not preforming as it should. This needs to be fixed but for now just play
safe and put proper bounds on the starting indexes