Interpret map into inserts, deletes & moves
The map is a helpful way-point but really need to be able to report on
where insert, deletes and moves have occurred. This update should start to
do that.

It only finds single row / columns, so that a contiguous range will get
reported multiple times rather that as one block (as it should be). I
think the fix is to run a secondary process to clean up translations that
could be grouped, for another time...