Allow distance measure to find non-exact matches
I wrote this to improve the matching behaviour by allowing the code to
find close matches. However, as currently coded, non-strict matches only
result in worse matching.

As a concept non-strict matching could be useful in the future so have
left the code changes in but with it set up so that we always perform a
strict search.

Based on my testing I'm thinking we need to do a full match of the entire
sheet, strictly first and then do a second pass with fuzzy matching
switched on to see if it can mop up the unmapped rows / cols. This is
definitely one to keep an eye on whilst doing more realistic testing