Memoise comment
Turns out that repeatedly finding a column or row from a sheet is a calc
bottleneck. Performance would be improved if we memoised the subsetted
rows and columns so that they're only looked up once.

On reflection I think this is best to attach to the CondensedFormula
object. We can start with a couple of null pointer lists for rows / cols.
Then as each row or col is called for, first check for a null pointer in
the lists or a stored object. If null pointer then get the row / col as
currently but then replace the null pointer with the got row / col object.
If we have an object then just return the row / col object without having to build
it again. Trading CPU for memory...