Memoise subsetted rows and columns
This was a performance bottleneck so am now storing the already found rows
and columns in a hash map. The hash map starts empty and only gets
populated as each row or column is needed for the first time. Thereafter a
row of column requested for a second or subsequent time will just return a
fast lookup from the hash map store