Improve coverage
Unit test coverage highlighted that I wasn't testing all the branches of
CompoundRange.equals(). Specifically I wasn't testing cases where the
count of cells in the object was the same but the individual cells
differed.

In doing the testing it uncovered another bug where I wasn't moving the
position index back to the start before looping through to look for cells
missing / present. This is now fixed