Move POI interaction into single class
POI is a serious bottleneck for performance and also messes up my unit
testing as I don't know how easily accommodate an external library.

For these reasons it's nice to be able to package up all the POI
interaction messiness into a single class that loads the file, can iterate
through the sheets of the workbook and for each sheet can spit out the
CondensedFormulae object that the rest of my code base uses.

The other benefit of structuring like this is that if, in the future, I
decide to write out the dependence on POI from the project I only need to
look at this one class. I suspect POI is seriously over-powered for my
needs and it seems to crash when opening big files (although that might
just be an environment memory issue)