Don't need to capture return object
This is a follow-up to an earlier commit. I've now been through the
code-base and I think this is the only instance that needs correcting.

Basically my IDE (NetBeans) was complaining that I never used the assigned
object. I have now realised I don't need to capture the returned object
and treat the method as though it returns void.