Translate From before doing diff
The FINAL piece in the puzzle! now we get a sensible diff because the rows
have been lined up first. After all that work we've finally got the
functionality I was originally hoping for.

Need to loop back and sort all out all those missing unit tests. Also need
to finish off the looping through sheets so the program works as a
stand-alone command line app.