Report cell values
It's helpful to be able to see the value in a cell, formatted exactly as
it appears in Excel. This is particularly the case for inputs that are
just standalone numbers but may be formatted into dates, percentages or
some other text.

Formula now stores the display text of the cell and report() reports it