Remove empty constructor
There's no need to allow creation of an empty CondensedFormulae object.
This is because the add method is private, so there's nothing to do with
an empty object. Further, even if the add method were public, we still
need to build a ref map after having added all cells. For these reasons we
want to control proper creation through the construction with a list of
formulae