Add unit tests for translations
This is just a first pass and is not complete.

As part of doing this I've written a small bit of VBA that can translate a
sheet into the list of Formulae that are required to create a
CondensedFormulae object in the vernacular of this project. Otherwise
writing out test cases was going to get a little tedious!

In writing these tests it has become apparent that the translation
detection where we have multiple shell matches isn't working as we would
hope. You can see this in the column translation tests I have set up,
which correspond to a2.xlsx being mapped to c2.xlsx in the resources
folder. Need to find a cold damp towel, put it over my head and refine the
translation detection logic.

The trouble with translations is that both A1 notation and R1C1 notation
formulae get updated so there's no simple way to find equivalent but
translated cells. Since A1 style gets disturbed less I think the vague
plan is to iterate through rows / columns looking for A1 notation matches.
As each successive translation is "found" the target A1 notation formulae
being looked for will have to be updated to account for the already found
translations. The complexity will be that a formula might refer right or
down (bad model design, I know) and those references may yet be translated
further, meaning you are looking for the wrong target A1 formula. I guess we'll
see how we get on...