Added new CompoundRange constructor
It was a bit of a hassle building up a CompoundRange cell by cell so can
now pass a (well-formatted!) cell reference string in and the constructor
will unbundle that into the individual cells and each one.

In doing this I found a bug that column Z was not being correctly written
out. It's another hang-over from na-ka-na's zero based conversion utility
functions that I originally borrowed