Add unit tests for SheetDiff
Last major piece of missing unit test now in place. The testing of this
class can, should and will be more rigorous but for now I get the code
coverage boosted.

I'm a little wary on being overly specific on the unit tests here as I'm
testing the written output of the program and I'm not 100% sure I've got
the right format yet: it feels way too verbose.

Also I'd like to try outputting in the form of Unified Output
http://ftp.gnu.org/old-gnu/Manuals/diffutils-2.8.1/html_mono/diff.html#Unified%20Format