Find cell references off sheet & workbook
Extended the CellRef object to additionally hold sheet and workbook
information. Now the formula parsing logic will find off-sheet references
and external links and include this information in the extended CellRef
object.

This re-write has been prompted by a crash when trying to translate an
off-sheet reference. In particular the external row reference was to a row
greater than the UsedRange of the current sheet, which caused an out of
bounds error as it tried to lookup the mapping for a row that didn't
exist. The solution is to note when references are off sheet or workbook
and to not translate them

As noted in the comments, we might try to consider all other sheets
translations and apply the relevant sheet's translations to off-sheet
references. This is definitely a future task as will be complicated