Don't interpret non-formulae
Found an issue with literal cells that had text in it that looked like a
cell reference. So now added an initial gate to check the cell formula
starts with an = sign, which is the hall-mark of a formula

Thinking about it we could still get caught out with cell references in
string literals e.g. ="Q1 to Q4 actauls:" & TEXT($Z$3,"#,##0"). IN the
above example only $Z$3 should be found as a cell reference but Q4 will
also get picked up. This still needs to be fixed