Pre-build rows and columns
Noticed that finding (sub-setting) rows and columns from a sheet was
killing a lot of time. Given we need to find every row and column in the
translation search anyway might as well bite the bullet and create them
up-front.

Seems to reduce run-time to under a third of it's previous time so a
pretty massive win