Update convert to R1C1 process
Another change driven by process efficiency. Noticed we were losing a lot
of time finding the cell ref a second time over in order to actually do
the replacement so have switched over to matcher's appendReplacement
method which builds an edited string in situ as you loop through the
search string.

This has a pretty massive effect on performance and code that used to take
3 mins on a 2Mb workbook a couple of days is now down to 8 secs. To be
fair 3 mins became 50 secs with the change to pre-can the rows and columns
of a sheet represented as CondensedFormulae object and the latter
reduction from 50 secs to 8 secs is the consequence of this change. We're
now getting into the territory of the code being performant enough to use